mostly norm fixed image file creation and a lot of little fixes and testing

+++ src/math/rt_math.c
@@ -40,7 +40,7 @@ float length_sq(t_v3 a)
 inline
 float square_root(float a)
 {
-	float result = (float)sqrt(a);
+	float result = sqrtf(a);
 
 	return (result);
 }
